Pad and join multiple strings within character vectors

Joins one or more character vectors as in str_c but with the character strings in each vector padded to make them equal length. The last vector is by default left unpadded but can optionally be padded too.

pad_c(..., pad_last = FALSE, sep = " ", collapse = NULL)

Arguments

...
one or more character vectors as for str_c.

pad_last
logical: pad the last vector in addition to the others or leave it without padding?

sep
character: separator between vector elements.

collapse
character: optional string used to combine input vectors into a single string.

Value

character: a single string constructed in a similar way as with str_c but where vectors are padded to a uniform width. If collapse is NULL (the default) the returned value is a character vector with length equal to the length of the first input vector. If collapse is non-NULL, the returned value is a character vector of length 1.

See also

Examples

vec1 <- c("Avocado:", "Blueberry:", "Coconut:", "Dill:")
vec2 <- c(rep("fruit", 3), "herb")
vec3 <- c("green", "purple", "white", "green")

pad_c(vec1, pad_last = TRUE)

#> [1] "Avocado:  " "Blueberry:" "Coconut:  " "Dill:     "

cat(pad_c(vec1, vec2, collapse = "\n"), "\n")

#> Avocado:   fruit
#> Blueberry: fruit
#> Coconut:   fruit
#> Dill:      herb 

cat(pad_c(vec1, vec2, vec3, collapse = "\n"), "\n")

#> Avocado:   fruit green
#> Blueberry: fruit purple
#> Coconut:   fruit white
#> Dill:      herb  green