Pluralize word

Converts a word into a plural (English).

pl(singular, num, irregular = NULL)

Arguments

singular
character: singular form of the word.

num
integer: number of items. num = 1 will give the singular form; num = 0 or num > 1 will give the plural form.

irregular
character: plural form of the word if it is irregular and needs to be explicitly supplied.

Value

character: singular or plural form of the word, depending on the value of num.

Details

This function would usually be used in a context where the number of items num is a variable and not known at run-time.

Examples

pl("shoe", 2)

#> [1] "shoes"

pl("shoe", 1)

#> [1] "shoe"

n <- 0
pl("thing", n)

#> [1] "things"

pl("sheep", 5)           # => incorrect (gives "sheeps")

#> [1] "sheeps"

pl("sheep", 5, "sheep")  # => correct   (gives "sheep")

#> [1] "sheep"