Set RNG seed

Sets the RNG (random number generator) seed.

set_init_seed(seed_val)

Arguments

seed_val
one of the following:

a single character value (which can be interpreted as an integer) to set the RNG;

the character literal "random" (to set a random seed);

NULL, to re-initialize the RNG as if no seed had yet been set (see 'Details' in set.seed); or

NA, in which case the function will return without any change to the seed.

Value

NULL, invisibly

See also

Examples

if (FALSE) {
set_init_seed(1)        # => set seed to a specific value
set_init_seed("random") # => set seed to a randomly-generated value
set_init_seed(NULL)     # => re-initialize seed as if none had yet been set
set_init_seed(NA)       # => return immediately without changing seed
}